Make some async api work on Scala.js and Native
This PR breaks binary compatibility in Scala.js and Scala Native but
doesn't touch binary compatibility for Scala JVM.
Since the methods removed weren't linking on JS/Native this can be
released as a patch release, until we find an API that works seamlessly in
all the platforms.
This deprecates Context.Impl.scheduler which cannot be implemented on
Native and JS and suggests to use scheduleMsg instead.